Add files via upload
Parallel_FSS.py is a parallelized  and optimized version of the old Finite Size Scaling program create by chase. I used Cython to optimize the routine. Additionally i used the parallel library in cython  which uses OpenMP to parallelize the for loops in our routine. You can simply import the module Parallel into your python file to have access to the optimized and parallelized FSS routine! 

Usage is as follows:

 Parallel.fssRoutine(float minDifference (initially value for the minimum value of fit for the routine),
                               FSS Lists (2D numpy array containing the data you want to collapse), 
                               pcSpace(1d numpy array that contains the range for the percolation threshold), 
                               nuSpace(1D numpy array that contains the range for the critical exponent Nu),
                               zetaSpace(1D numpy array that contains the range for the correlation length), 
                               rho(1D numpy array that contains the range for the critical exponent Rho))